requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0

# Machine learning libraries
scikit-learn>=1.2.0
xgboost>=1.7.0
lightgbm>=3.3.0

# Optimization
optuna>=3.0.0

# Utility libraries
joblib>=1.2.0

# Date/time handling
holidays>=0.25

# Database connectivity
pyodbc>=4.0.34
sqlalchemy>=1.4.0

# Environment variable management
python-dotenv>=0.19.0

# Progress bars and logging (standard library alternatives available)
# logging - built-in
# typing - built-in
# os - built-in
# datetime - built-in
# warnings - built-in